fix(integ-tests): Uint8Arrays are not decoded properly  (#27009)
AWS SDK v3 has changed some data types to Uint8Array. If a value of this type was provided by the user, it would not be decoded properly back to a Uint8Array before making the call.

Recognize when there is an encoded Uint8Array and decode it back properly.

Closes https://github.com/aws/aws-cdk/issues/26798.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*