fix(events-targets): AwsApi is still using Node 16 (#27002)
Migrate the `aws-api-handler` code to use SDKv3 and move it into `custom-resource-handlers`.
Updates the Construct config to use the new code and Node18.
To verify functionality, the existing integration test has been extended to include a new rule and target that can be asserted on.

Includes a config change to the `custom-resource-handlers` package to run test from TS.

Closes #26998

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*