fix(events-targets): AwsApi fails when SDKv3 Command expects Uint8Array (#27034)
In #27002 we have migrated the AwsApi Construct to SDKv3.
However some SDKv3 Commands expect a `Uint8Array` in places where SDKv2 previously would accept a `string`.
This is a problem because handler input is passed directly into the SDK calls and the input can only contain JSON data types.
To solves this, we keep a list of SDK actions and parameters that should be coerced to `Uint8Array`.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*