chore: use error name string in handlers
Use the `error.name` field to disambiguate errors in custom resource
handlers. Instanceof has errors if the service team has not correctly
modeled their errors which leads to ambiguity when detecting and
handling specific errors returned from the service.

`.name` usage is reported to be backwards compatible to sdk v2.