fix(codedeploy): CustomLambdaDeploymentConfig is broken (#27087)
Since the conversion of SDKv2 to SDKv3 of the `AwsCustomResource`, this custom resource was broken with the following error:

```
SerializationException: STRING_VALUE can not be converted to an Integer
```

The reason is that SDKv2 used to accept strings and coerce them to numbers wherever necessary, but SDKv3 does not do this anymore.

Update the Custom Resource usage to no longer stringify the numbers.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*