chore(msk): integ test for zookeeper connect custom resource (#27102)
Adding an integ test for this custom resource with our new rule that each custom resource needs at least one integ test. There is nothing to assert here, since the custom resource calls an AWS api to return a string. Asserting it with the same api call seems redundant. Instead, I'm going with supplying a cfnOutput to the console, which shows the string is zookeeper related.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*