fix(triggers): errors are shown as list of bytes (#27119)
With Triggers a user can invoke a Lambda Function during every stack deployment.

When we migrated the handler code to SDKv3, we overlooked that the return payload from the invoked lambda function is now a `Uint8Array`, and not a string anymore.
In the error case, we attempt to `JSON.parse()` the error payload which then resulted in a list of bytes being returned to the user.

Updated the handler test case to use correct mocks, including returning a payload that is a Uint8Array.

Closes #27113

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*