docs(codepipeline): add note about repoString param resolvability at runtime (#27109)
This PR addresses an issue encountered when setting up an AWS CDK pipeline where unresolved token values for the `repoString` parameter led to a runtime error. The updated docstring now explicitly mentions that the `repoString` must be resolvable at runtime and that unresolved tokens are intended to be a runtime error

Closes #27100.

Thanks to @peterwoodworth for the quick assessment and the proposed solution to the problem.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*