fix(custom-resources): does not accept strings where numbers are expected, skips recursive types (#27112)
AWS SDK v3 strictly validates numeric types, as opposed to v2, which allowed, for example, "123" to be passed where an integer was expected.

This script adds all attributes of a numeric type ("byte", "short", "integer", "long", "bigInteger", "float", "double" and "bigDecimal") to a new type coercion map. It also compacts the map by trimming each part of the path to its unique prefix among all the paths in its operation.

We also change the data structure from a list of paths (forming a tree) to a graph/state machine. This is necessary in order to support self-recursive types, which can loop arbitrarily and therefore cannot be represented in fixed length prefix lists.


See https://github.com/aws/aws-sdk-js-v3/issues/5198


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*