fix(synthetics): synth-time failure for canary assets in nested stages (#27167)
Fixes #27089.

Canaries inside stages would fail at synth time after #26291 was introduced. This was due to a misunderstanding of `assetOutdir`. It turns out that `this.asset.assetPath` is a relative path of `Stage.of(scope).outdir`, not of `Stage.of(scope).assetOutdir`. These two out directories would be equivalent at the root level, but if there were nested stages, `assetOutdir` would equal `../${outdir}`.

This diagram may help:

```
CLOUD ASSEMBLY ROOT
  +-- asset.12345abcdef/
  +-- assembly-Stage
    +-- MyStack.template.json
    +-- MyStack.assets.json <- will contain { "path": "../asset.12345abcdef" }
```


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*