fix(cli): equals sign in a tag value is dropped (#27130)
When overriding tags from the CLI the tag key-value string is split by an equal sign and the left and right parts are used for key and value respectively. The documentation for the [String.prototype.split()](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/split) states that the second parameter sets a limit of the returned values after execution:
> If provided, splits the string at each occurrence of the specified separator, but stops when limit entries have been placed in the array. Any leftover text is not included in the array at all.

So basically if we have `foo=bar=test` and we apply a split with a limit of 2 (`"foo=bar=test".split("=", 2)`) we will get the following array `["foo", "bar"]` instead of the expected one`["foo", "bar=test"]`. 

There has been the same problem with the context values and since it is the same problem the solution is also the same: #5773 

With this fix the `context` and `tags` get the same behaviour when parsing the passed values.

Closes #21003 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*