feat(ecr): make `validateRepositoryName` errors human readable (#27186)
When a user specifies an invalid repository name, the error message will now describe why the name is invalid instead of providing the regex used to validate the name. The message comes from the console when defining a new repository. The docstring of `repositoryName` now reflects this and mirrors the docstring of the underlying CFN resource.

Closes #26715.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*