fix(apigateway): CORS response has unexpected space
Because of the way Velocity templates are parsed, the space
just before `#end` in the original template was emitted literally
instead of swallowed.

This caused the `OPTIONS` response to contain a single space (`" "`),
which makes some validation tools reject the response as invalid.

Rewrite the Velocity template to a syntax we know doesn't cause
that space to be emitted.