fix(rds): prevent rendering deprecated credentials when creating a database cluster from a snapshot (under feature flag) (#27174)
This PR fixes a bug where an extra database secret is being generated when an RDS database cluster is being created from a snapshot.

On the `DatabaseClusterFromSnapshotProps` interface, we deprecated the `credentials` property and, at the same, introduced `snapshotCredentials` as the recommended replacement. However, the default behavior associated with the `credentials` property was not removed as doing so would introduce a breaking change for some users as detailed in this [PR](https://github.com/aws/aws-cdk/pull/20777). As a result, users just using the recommended `snapshotCredentials` property to create a new RDS database cluster are seeing an extra, unwanted secret being created.

Closes #23815

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*