fix(appconfig): allow multiple environment monitor roles to be created (#27243)
Currently, when adding multiple monitors under an environment, users are seeing the following error

```
Error: There is already a Construct with name 'Role' in Environment [MyEnvironment]
```

Before the logical id for any monitor role that is created was just `Role`. In this change, we use the monitor index in the logical id so there aren't any repeats in logical id's. This is also why there is a change in the integ test snapshot, since the logical id was refactored.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*