fix(batch): make `instanceType` optional on multinode jobs (#27223)
Make `instanceType` optional. It used to be required. It will default to a batch optimal instance type.

The breaking change is for `MultiNodeJobDefinition.instanceType`, which is now optional. This is not truly breaking because everyone has this property explicitly set now, and will continue to have it set after this change. 

Closes #27185.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*