build(awslint): Add RuleFilterSet class to cache include/exclude rules. (#27319)
Incremental fix to awslint performance. The `shouldEvaluate()` method is invoked for every context of every rule of every linter. The rule filter matching that was called from this method was iterating over the include/exclude list each time, and since we now have 2k+ rule filters, this was taking a long time( 55s for `aws-cdk-lib`).

This PR caches those rule filters and causes most lookups to be O(1), taking the evaluation time from 55s to 18ms. One minor limitation in the cache is that you can no longer use a 'startsWith' asterisk in the code portion of the filter, though you can still use an 'anyString' wildcard for the whole. This is explained better in the the comments:

```
  /**
   * Rule filter format is code:scope e.g. 'docs-public-apis:aws-cdk-lib.TagType'.
   * Wildcards can be used in the following ways:
   *
   * 1. Match any code e.g.: '*:aws-cdk-lib.TagType'
   * 2. Match any scope e.g.: 'docs-public-apis:*'
   * 3. Match any scope prefix e.g.: 'docs-public-apis:aws-cdk-lib.Tag*'
   */
```

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*