fix(autoscaling): ssmSessionPermissions does not add managed policy to role specified in launch template (#27220)
When we set ssmSessionPermissions to true, CDK adds `AmazonSSMManagedInstanceCore` managed policy to role specified in AutoScalingGroup construct. However, if the role specified in not AutoScalingGroup construct's prop but launch template, ssmSessionPermissions has no effect.

This PR solves the issue by adding the managed policy even when role is specified in launch template.

Closes #25904

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*