chore: align-version script drops caret from dependencies it rewrites
The `align-version` script is supposed to rewrite `0.0.0` -> `2.99.0`,
and also to rewrite `^0.0.0` -> `^2.99.0`. It didn't do the latter,
because of a variable shadowing mistake. That mistake happened to come
out correctly often enough that we didn't notice it.

Fix that bug.

Also add a linting rule that says that all peerDependencies should
have a `^` dependency on `aws-cdk-lib`.