chore(synthetics): restore synthetics alpha module (#27394)
Synthetics was stabilized in #27305 and the alpha module was removed. Due to a bug in `align-versions.sh`, the last published `@aws-cdk/aws-synthetics-alpha` module relied on a peer dependency of `aws-cdk-lib@2.99.0` instead of `aws-cdk-lib@^2.99.0` (with the caret). This makes `@aws-cdk/aws-synthetics-alpha incompatible with later versions of aws-cdk-lib, and is unintentional.

#27376 fixes the bug in `align-versions.sh`, and this PR re-introduces `@aws-cdk/aws-synthetics-alpha` for one last version of CDK (v2.100.0) before we remove it again. This should have the effect of `@aws-cdk/aws-synthetics-alpha@2.100.0` being the last version of the alpha module, but still compatible with later versions of `aws-cdk-lib` by having the peer dependency `aws-cdk-lib@^2.100.0`. 

We still want users to migrate to the stable synthetics module at `aws-cdk-lib/aws-synthetics`, but we don't want to outright break existing users of the experimental module either.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*