fix(logs): remove invalid data protection data identifier (#27388)
### Notes 
`DataIdentifier.PHONENUMBER` is not a valid identifier (this can be verified in AWS Console), this should be removed to avoid confusion. Adding this identifier will cause `An error occurred (InvalidParameterException) when calling the PutDataProtectionPolicy operation: arn:aws:dataprotection::aws:data-identifier/PhoneNumber is not a valid Data Identifier`.

Note: I am an SDE on CloudWatch Logs, this was brought up by a user, please message kvcg@ internally for questions regarding this PR. 

### Testing
Unit tests passing in `aws-logs/test/loggroup.test.ts`.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*