fix(glue-alpha): prefix validation logic is incorrect (#27472)
There was logic in `aws-glue-alpha.Job` that expected `SparkUIProps.prefix` to be specified in the form `/prefix-name`, instead of the more conventional `prefix-name/`. This fix addresses the validation and handling of this prefix field to support the correct format. 


BREAKING CHANGE: `SparkUIProps.prefix` strings in the original `/prefix-name` format will now result in a validation error. To retain the same behavior, prefixes must be changed to the new `prefix-name/` format.

Closes #27396.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*