chore(synthetics-alpha): remove deprecated flag from package.json (#27505)
This is just a hunch, but [this](https://stackoverflow.com/questions/67849132/verdaccio-publish-new-package) stack overflow answer claims that the `deprecated` tag equals malformed package.json and that is the reason for the `no such package available` error I am seeing in testing.

We've added this `deprecated` tag before and it's been fine and as far as I know we don't use Verdaccio when we release but only when we test (not entirely sure of the truthfulness of that statement). However, this line is not important enough to me anyway, so I'm removing it.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*