chore(@aws-cdk-testing/cli-integ): add sinon dep
Add an explicit dependency on a version of sinon that we depend on in
other packages. This version is what is used by `aws-cdk-lib` currently.
This fixes an issue that occurs currently when installing
@aws-cdk-testing/cli-integ in a directory that is not a git repository
as described [in this issue](https://github.com/sinonjs/sinon/issues/2557).

This doesn't occur during local development because yarn dedupes sinon
to a version that intersects with the peerDep range taken by
`ts-mock-imports` and versions depended on by other packages, such as
`aws-cdk-lib`.