fix(pipelines): using the same source (s3, connection, or codecommit) in a pipeline causes duplicate id error (#27602)
Differentiates between sources of the same repository by appending the branch name onto the node id and input/output artifacts. This avoids the duplicate id errors for different branches of the same repository, as well as validating that each source is a unique repository & branch combination.

The only change to the CFN template is the input & output artifacts, but since these are not stateful resources, they can be modified without breaking changes. The artifacts are also updates in tandem, so the pipeline source behavior will stay the same.

This change impacts these `CodePipelineSource`s:
- `s3()` - `objectKey` appended
- `connection()`
- `codeCommit()`

Does not change `ecr()` behavior as there is no notion of folders or branches.
Does not change `github()` behavior as doing so would cause destructive changes of webhooks.

Closes #23916 and #19875.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*