fix(ec2): naming collisions when using `ec2.InitFile.fromAsset()` on multiple instances in the same stack (#27468)
Closes #16891 

If a user creates more than one EC2 instance in the same stack and defines an InitConfig for each instance using `ec2.InitFile.fromAsset()`, the user will get an error if they pass in the same `targetFilePath` or `targetDirectory`. This bug is due to [how the asset is uploaded to S3](https://github.com/aws/aws-cdk/blob/main/packages/aws-cdk-lib/aws-ec2/lib/cfn-init-elements.ts#L427).

This PR fixes this issue by adding a hash to the `targetFileName` before being uploaded as an s3 asset.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*