chore(prlint): apply needs-maintainer-review label when PR is assigned
PRs can only be assigned by members with write access. If a PR has been
assigned to someone, remove the `needs-community-review` label (since it
implies a maintainer is already working on that PR) and apply
the `needs-maintainer-review` label. In order to facilitate this, the
`pull_request_target` event should also be run when a PR is
assigned/unassigned.