feat(eks): support updateConfig properties for managed node groups (#27593)
This is a duplicate of [PR 27320](https://github.com/aws/aws-cdk/pull/27320) but on a new branch in my repo.

This change enables nodegroups to configure updateConfig properties which will allow nodegroups to upgrade more than 1 instance at a time.  The default value of 1 results in slow deployment times which prevented my service pipelines from quickly delivering fixes to my customers.  This change allows each nodegroup to specify a maximum count of instances, or percent, which can be upgraded in parallel.