chore: add more details to CONTRIBUTING (#27703)
Add more details to our contributing guidelines about what we look for in contributions to aws/aws-cdk and what may be better suited to third party packages. If there are other details and guidelines that we should enumerate up front we should add those here as well.

Added a fancy little diagram cause they are fun and I'm the best.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*