chore(cloudfront): create unit tests to mock edge function handler (#27768)
This PR adds three unit tests that mock functionality found in the Lambda handler for edge functions. Specifically, this PR is adding one unit test for a create event, one unit test for an update event, and one unit test for a delete event to verify that the handler only executes for create and update events.

Closes #27724

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*